Add Delay sections
