+++ apps/robots-generator/src/components/Delays/index.ts
@@ -1,0 +1,3 @@
+import Delays from "./Delays";
+
+export default Delays;
